Improved polygon rendering function. All pixels rendered on the boundary (i.e. pixels horizontally or vertically adjacent to a pixel that is not part of the polygon) are now correctly excluded when testing for overlapping pixels. This means the server no longer needs the clear-previous-segment-before-drawing-next-segment hack to prevent overlap from being signaled when adjacent segments are drawn.
However, there is now a different hack in the draw_poly function, which is
required to ensure that segments (a,b,c,d) and (d,c,e,f) which share the edge
(d,f) only overlap in boundary points. This is not the case by default because
clipping is asymmetrical: the line (c,d) contains different pixels than the
line (d,c). This should be fixed later.