Ensured that in a head-on collision, both players get a point.
If player x dies in turn t, player y gets a point if he died in turn t-1 and
y > x, thereby removing the advantage of the earlier player in a head-on
collision. This gives the later player a slight advantage in other situations
(e.g. if player 1 survives exactly one turn longer than player 2, the latter
 still gets a point). The interval between turns is small enough for this to
be unnoticable.