Made it harder to score hole points without actually going through the hole.
Previously, touching the hole was enough to score a point, as holes had the
same thickness as regular lines. This was changed so holes are actually much
less wide, so a player must touch the center of the hole in order to score a
point. (Although scoring without going through is still possible, it is much
harder if not impossible in practice.)